Fix shaping default bandwidth
The current code uses cfg.DefaultBandwidth when addTargetClass() which
is not right at first sight.

The other thing is that HTB seems to have a default built-in class where
all unclassified packets will go into.  This can be checked with value
of direct_packets_stats in output of "tc qdisc show dev eth0", and as
such we need to explicitly set the default class to 10:1 to handle this